+++ package.json
@@ -30,6 +30,7 @@
     "pug": "^2.0.0-beta6",
     "readline-sync": "^1.4.5",
     "rimraf": "^2.6.2",
+    "sendmail": "^1.2.0",
     "sweetalert": "^1.1.3",
     "uuid": "^3.1.0"
   }
