+++ public/css/alert.css
@@ -1,4 +1,4 @@
-#modalAlert {
+#alert {
 	background-color: white;
 	max-width: 500px;
 	max-height:375px;
