+++ views/feedback.pug
@@ -1,0 +1,28 @@
+block style
+  link(rel='stylesheet' href='/css/feedback.css')
+
+block content
+  #feedback.modal
+    .modal-content
+      h3 Feedback
+
+      .row
+        form.col.s12
+          .row
+            .input-field.col.s12
+              i.material-icons.prefix label_outline
+              input#subject(placeholder='Subject', type='text')
+              label(for='subject') Subject
+          .row
+            .input-field.col.s12
+              i.material-icons.prefix mode_edit
+              textarea#feedbackArea.materialize-textarea(placeholder='Message', type='text')
+              label(for='feedbackArea') Message
+          .row
+            .input-field.col.s6
+              button.modal-action.modal-close.btn.btn-info Cancel
+            .input-field.col.s6
+              button.modal-action.modal-close.btn.btn-info(type='button') Submit
+
+block scripts
+  script(src='/js/feedback.js')
