+++ views/navbar.pug
@@ -44,4 +44,4 @@ ul#mobile.side-nav
   li#nav-logout
     a.navbarLinkHidden(href='/logout') Logout
 
-include modal-feedback
+include feedback
