+++ views/question-view.pug
@@ -57,7 +57,7 @@ block content
         .col.s12
           if answered || isAdmin()
             #discussion
-  include modal-alert
+  include alert
 
 block scripts
   script(src='/js/common.js')
